+++ src/Raven.CodeAnalysis/Syntax/InternalSyntax/Parser/Parsers/NamespaceParser.cs
@@ -297,6 +297,15 @@ internal class NamespaceDeclarationParser : SyntaxParser
             AddMemberDeclarationWithSeparatorValidation(macroDeclaration);
             order = MemberOrder.Members;
         }
+        else if (new MacroInvocationSyntaxParser(this).IsDeclarationInvocationStart())
+        {
+            var declaration = new MacroInvocationSyntaxParser(this).ParseDeclaration(
+                SyntaxList.Empty,
+                SyntaxList.Empty);
+
+            AddMemberDeclarationWithSeparatorValidation(declaration);
+            order = MemberOrder.Members;
+        }
         else if (nextToken.IsKind(SyntaxKind.ConstKeyword) ||
                  nextToken.IsKind(SyntaxKind.EnumKeyword) ||
                  nextToken.IsKind(SyntaxKind.UnionKeyword) ||
@@ -328,6 +337,16 @@ internal class NamespaceDeclarationParser : SyntaxParser
                 var macroDeclaration = macroParser.Parse(attributeLists, modifiers);
 
                 AddMemberDeclarationWithSeparatorValidation(macroDeclaration);
+                order = MemberOrder.Members;
+                return;
+            }
+
+            var declarationMacroParser = new MacroInvocationSyntaxParser(this);
+            if (declarationMacroParser.IsDeclarationInvocationStart())
+            {
+                var declaration = declarationMacroParser.ParseDeclaration(attributeLists, modifiers);
+
+                AddMemberDeclarationWithSeparatorValidation(declaration);
                 order = MemberOrder.Members;
                 return;
             }